persist: move PersistConfig into a new module
It's gotten complex enough to warrant it and the dynamic config stuff
will only continue down that path.
