+++ src/environmentd/tests/util.rs
@@ -105,7 +105,8 @@ use mz_ore::retry::Retry;
 use mz_ore::task;
 use mz_ore::tracing::TracingHandle;
 use mz_persist_client::cache::PersistClientCache;
-use mz_persist_client::{PersistConfig, PersistLocation};
+use mz_persist_client::cfg::PersistConfig;
+use mz_persist_client::PersistLocation;
 use mz_secrets::SecretsController;
 use mz_sql::catalog::EnvironmentId;
 use mz_stash::StashFactory;
