+++ src/persist-client/benches/benches.rs
@@ -93,8 +93,9 @@ use mz_persist::postgres::{PostgresConsensus, PostgresConsensusConfig};
 use mz_persist::s3::{S3Blob, S3BlobConfig};
 use mz_persist::workload::DataGenerator;
 use mz_persist_client::async_runtime::CpuHeavyRuntime;
+use mz_persist_client::cfg::PersistConfig;
 use mz_persist_client::write::WriteHandle;
-use mz_persist_client::{PersistClient, PersistConfig};
+use mz_persist_client::PersistClient;
 use mz_persist_types::Codec64;
 
 // The "plumbing" and "porcelain" names are from git [1]. Our "plumbing"
