+++ src/sqllogictest/src/runner.rs
@@ -69,7 +69,8 @@ use mz_ore::task;
 use mz_ore::thread::{JoinHandleExt, JoinOnDropHandle};
 use mz_ore::tracing::TracingHandle;
 use mz_persist_client::cache::PersistClientCache;
-use mz_persist_client::{PersistConfig, PersistLocation};
+use mz_persist_client::cfg::PersistConfig;
+use mz_persist_client::PersistLocation;
 use mz_pgrepr::{oid, Interval, Jsonb, Numeric, UInt2, UInt4, UInt8, Value};
 use mz_repr::adt::date::Date;
 use mz_repr::adt::numeric;
